requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
anthropic>=0.40.0
instructor>=1.4.3
langgraph>=0.2.28
pdfplumber>=0.11.4
pydantic>=2.9.2
python-multipart>=0.0.12
python-dotenv>=1.0.1
pytest>=8.3.3
pytest-mock>=3.14.0
